+++ tests/src/FlexTest/Data/AbstractObjectTest/Object.php
@@ -10,7 +10,6 @@ use Flex\Data\AbstractObject;
  */
 class Object extends AbstractObject
 {
-
     /**
      * @return string
      */